% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/XSim_mod.R
\name{XSim.mod}
\alias{XSim.mod}
\title{Bi-clustering using Co-Similarity width modification}
\usage{
XSim.mod(x, y, itr = 4)
}
\arguments{
\item{x}{matrix or dataframe of predictors, of dimension n*p; each row is an observation vector.}

\item{y}{response variable (1 or 2)}

\item{itr}{number of iterations.}
}
\value{
Return a list of objects.
}
\description{
A modification of co-similarity by Hussain for Gene Expression Data ( See in the reference)
}
\examples{
library(PDI2015)
XSim.mod(ColonCancer[, -1], ColonCancer[,1], itr = 4)
}
\references{
Hussain S.F. \emph{Bi-Clustering Gene Expression Data Using Co-Similarity}, 7th International Conference on Advanced Data Mining and Applications (ADMA), 16-19th Dec. 2011, Beijing, China.
}

