% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/XSim2008.R
\name{XSim2008}
\alias{XSim2008}
\title{Bi-clustering using Co-Similarity (version in 2008)}
\usage{
XSim2008(x, y, itr = 4)
}
\arguments{
\item{x}{matrix or dataframe of predictors, of dimension n*p; each row is an observation vector.}

\item{y}{response variable (1 or 2)}

\item{itr}{number of iterations.}
}
\value{
Return a list of objects.
}
\description{
The χ-Sim algorithm is a co-similarity based approach which builds on the idea of simultaneously generating the similarity matrices SR (between documents) and SC (between words), each of them built on the basis of the other. This method is to implement the algorithm X-SIM in 2008 of Gilles Bission et Feward Sye Hussain. See in the Reference.
}
\examples{
library(PDI2015)
# Test with Colon Cancer dataset
res = XSim2008(ColonCancer[, -1], ColonCancer[,1], itr = 4)

# Test News Groups dataset
res1 = XSim2008(NG20SMI$M2[[1]],  NG20SMI$M2[["class"]],  itr = 4)
res2 = XSim2008(NG20SMI$M2[[2]],  NG20SMI$M2[["class"]],  itr = 4)
res3 = XSim2008(NG20SMI$M5[[1]],  NG20SMI$M5[["class"]],  itr = 4)
res4 = XSim2008(NG20SMI$M10[[1]], NG20SMI$M10[["class"]], itr = 4)
res5 = XSim2008(NG20SMI$NG1[[1]], NG20SMI$NG1[["class"]], itr = 4)
res6 = XSim2008(NG20SMI$NG2[[1]], NG20SMI$NG2[["class"]], itr = 4)
res7 = XSim2008(NG20SMI$NG3[[1]], NG20SMI$NG3[["class"]], itr = 4)
}
\references{
Bisson G, Hussain F. \emph{X-sim - A new similarity measure for the co-clustering task}. 7th IEEE International Conference on Machine Learning and Applications (ICMLA), 11-13th Dec. 2008, San Diego, United States.
}

