% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/XSim2015.R
\name{XSim2015}
\alias{XSim2015}
\title{Bi-clustering using Co-Similarity (version in 2015)}
\usage{
XSim2015(x, y, itr = 2)
}
\arguments{
\item{x}{matrix or dataframe of predictors, of dimension n*p; each row is an observation vector.}

\item{y}{response variable (1 or 2)}

\item{itr}{number of iterations.}
}
\value{
Return a list of objects.
}
\description{
Bi-clustering Biological Dataset using Co-similarity by Sang & Saad.
}
\examples{
library(PDI2015)
# Test with Colon Cancer dataset
res = XSim2015(ColonCancer[, -1], ColonCancer[,1], itr = 4)

# Test News Groups dataset
res1 = XSim2015(NG20SMI$M2[[1]],  NG20SMI$M2[["class"]],  itr = 4)
res2 = XSim2015(NG20SMI$M2[[2]],  NG20SMI$M2[["class"]],  itr = 4)
res3 = XSim2015(NG20SMI$M5[[1]],  NG20SMI$M5[["class"]],  itr = 4)
res4 = XSim2015(NG20SMI$M10[[1]], NG20SMI$M10[["class"]], itr = 4)
res5 = XSim2015(NG20SMI$NG1[[1]], NG20SMI$NG1[["class"]], itr = 4)
res6 = XSim2015(NG20SMI$NG2[[1]], NG20SMI$NG2[["class"]], itr = 4)
res7 = XSim2015(NG20SMI$NG3[[1]], NG20SMI$NG3[["class"]], itr = 4)
}

